Parameter update when executing clickCallback
When there are more than one entries returned the click call back is called with 2 parameters.  When 1 record is returned the clickCallback() is called with 1 parameter.  This caused inconsistencies in the use of clickCallback custom code.  

Code was updated to include a null value for 1st parameter when clickCallback() if there was only one record returned in the search.